Fix number of attempts and output redirection in cluster
Maximal number of attempts to run a container in cluster was previously set to 1 for testing.
aucont_cluster used to redirect stderr and stdout to different files, whilst the specification says that the streams should be redirected to the same file
Also changed the default stderr redirection. Now it's redirected to stdout instead of a separate file by default.